pyproject.toml
[tool.poetry]
name = "skctl"
version = "0.1.0"
description = "skctl is the Sharedkube's CLI tool"
authors = ["Patryk Kobielak <patryk.kobielak@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
click = "^8.1.7"
requests = "^2.32.3"
pyyaml = "^6.0.2"
tabulate = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
skctl = "skctl.skctl:cli"
